Split transaction type coverage E2E suite
